Cleanup: move more process-related code together
Incidentally, move some low-level RISC-V-specific code to riscv.c.
